Collect all code related to file IO into the nchg_io lib
