+++ src/nchg/cartesian_product/cartesian_product.cpp
@@ -50,9 +50,9 @@ NCHG_DISABLE_WARNING_POP
 #include <vector>
 
 #include "nchg/hash.hpp"
+#include "nchg/text.hpp"
 #include "nchg/tools/common.hpp"
 #include "nchg/tools/config.hpp"
-#include "nchg/tools/io.hpp"
 #include "nchg/tools/tools.hpp"
 
 namespace nchg {
