+++ src/nchg/expected/expected.cpp
@@ -38,9 +38,9 @@ NCHG_DISABLE_WARNING_POP
 #include "nchg/common.hpp"
 #include "nchg/concepts.hpp"
 #include "nchg/expected_values.hpp"
+#include "nchg/text.hpp"
 #include "nchg/tools/common.hpp"
 #include "nchg/tools/config.hpp"
-#include "nchg/tools/io.hpp"
 #include "nchg/tools/tools.hpp"
 
 namespace nchg {
