+++ src/nchg/merge/merge.cpp
@@ -43,12 +43,12 @@ NCHG_DISABLE_WARNING_POP
 #include <utility>
 #include <vector>
 
-#include "nchg/common.hpp"
 #include "nchg/k_merger.hpp"
 #include "nchg/nchg.hpp"
+#include "nchg/parquet_stats_file.hpp"
+#include "nchg/record_batch_builder.hpp"
 #include "nchg/tools/common.hpp"
 #include "nchg/tools/config.hpp"
-#include "nchg/tools/io.hpp"
 #include "nchg/tools/tools.hpp"
 
 namespace nchg {
